+++ R/query.R
@@ -134,14 +134,15 @@ TrajMeanVelocity <- function(trj, startIndex = 1, endIndex = nrow(trj)) {
 #' Calculates the step angles (in radians) of each segment, either relative to
 #' the previous segment or relative to the specified compass direction.
 #'
-#' Note that since turning angles are circular quantities, i.e. 360° == 0°, it is
-#' incorrect to treat them as linear quantities. In particular, do not calculate
-#' arithmetic means or standard deviations of turning angles. See Batschelet,
-#' (1981) for a detailed explanation and techniques for dealing with circular
-#' quantities.
+#' Note that since turning angles are circular quantities, i.e. 360° == 0°, it
+#' is incorrect to treat them as linear quantities. In particular, do not
+#' calculate arithmetic means or standard deviations of turning angles. See
+#' Batschelet, (1981) for a detailed explanation and techniques for dealing with
+#' circular quantities.
 #'
 #' @param trj the trajectory whose whose angles are to be calculated.
-#' @param lag Angles between every lag'th segment are calculated.
+#' @param lag Angles between every lag'th segment are calculated. Only applies
+#'   to non-directed walks, i.e. \code{compass.direction} is \code{NULL}.
 #' @param compass.direction If not \code{NULL}, step angles are calculated
 #'   relative to this angle (in radians), otherwise they are calculated relative
 #'   to the previous step angle.